[#30][fix] cookie 설정 및 feign 설정 변경 cookie 도메인 설정 logout 로직 유효성 검사 변경

+++ src/main/java/f4/auth/domain/auth/service/impl/AuthServiceImpl.java
@@ -87,6 +87,10 @@ public class AuthServiceImpl implements AuthService {
     String email = claims.getSubject();
     Date expired = claims.getExpiration();
 
+    if (!redisService.hasBlackList(accessToken) && expired.after(new Date())) {
+      throw new CustomException(CustomErrorCode.ALREADY_LOGOUT_USER);
+    }
+
     redisService.deleteData(email);
     redisService.setBlackList(accessToken, Duration.ofMillis(getExpiration(expired)));
   }
